// src/config/onedosh/onedosh.config.interface.ts
export enum OneDoshSupportedCryptoCurrencies {
  USDC = 'USDC',
  USDT = 'USDT',
}

export enum OneDoshSupportedCryptoNetworks {
  SOL = 'SOL',
  TRON = 'TRON',
  ETH = 'ETH',
  BTC = 'BTC',
  ERC20 = 'ERC20',
  BASECHAIN = 'BASECHAIN',
}

export enum OneDoshSupportedCountries {
  US = 'US',
  NG = 'NG',
}